chore!: drop support for Node.js < 18.20.4
Node.js v16 dropped maintenance support in late 2023, v14 was years earlier, and many package maintainers have since moved on. To keep this formatter relevant and secure, we are aligning with the end of major support windows. The next expected change is in January 2025, when v18 will drop support, and we will move to v20.

BREAKING CHANGE: drop support for Node.js < 18.20.4